tests: split up
Also go a little further on XML.
